Make slope rating an int. Add db migration. Format course and slope ratings so they display properly.

+++ src/GolfBag/Entities/TeeBox.cs
@@ -10,7 +10,7 @@ namespace GolfBag.Entities
 
         public string Name { get; set; }
 
-        public decimal SlopeRating { get; set; }
+        public int SlopeRating { get; set; }
 
         public decimal CourseRating { get; set; }
